+++ xraylabtool/interfaces/completion.py
@@ -69,6 +69,7 @@ _xraylabtool_install_completion_complete() { _xraylabtool_complete; }
 # SiO2 Si Al2O3 Fe2O3 CaO MgO Na2O K2O TiO2 P2O5
 
 complete -F _xraylabtool_complete xraylabtool
+complete -F _xraylabtool_complete xtool
 """